# hdms-backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: HDMS API
  description: >
    Hito Device Management System. Source of truth for the HTTP contract —
    see docs/06-api-contract.md. Go server interfaces are generated with
    oapi-codegen; the TypeScript client is generated with openapi-ts.
    Neither is hand-edited.
  version: "0.1.0"
servers:
  - url: /v1
security:
  - adminSession: []
paths:
...
  /auth/me:
    get:
      operationId: getCurrentAdmin
      summary: The identity of the currently authenticated admin.
      tags: [auth]
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Admin"
        default:
          description: Not authenticated.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"

  # --- Catalog --------------------------------------------------------------
...
  /categories:
    get:
      operationId: listCategories
      summary: List device categories.
      tags: [categories]
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryList"
        default:
          $ref: "#/components/responses/ProblemResponse"
    post:
      operationId: createCategory
      summary: Create a device category.
      tags: [categories]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryRequest"
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        default:
          $ref: "#/components/responses/ProblemResponse"
  /categories/{id}:
    patch:
      operationId: updateCategory
      summary: Edit a device category.
      tags: [categories]
      parameters:
        - $ref: "#/components/parameters/IDParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCategoryRequest"
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        default:
          $ref: "#/components/responses/ProblemResponse"

  # --- Identity ---------------------------------------------------------
...
  /credentials:
    get:
      operationId: listCredentialsBySubject
      summary: List every credential ever issued to one subject, most recent first.
      tags: [credentials]
      parameters:
        - name: subjectType
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/SubjectType"
        - name: subjectId
          in: query
          required: true
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/CredentialStatus"
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CredentialList"
        default:
          $ref: "#/components/responses/ProblemResponse"
    post:
      operationId: issueCredential
      summary: Mint a new credential. The plaintext token is returned exactly once.
      tags: [credentials]
      x-sensitive: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IssueCredentialRequest"
      responses:
        "201":
          description: Created. Never cached.
          headers:
            Cache-Control:
              schema:
                type: string
                enum: [no-store]
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IssuedCredential"
        default:
          $ref: "#/components/responses/ProblemResponse"
  /credentials/blank-batch:
    post:
      operationId: issueBlankBatch
      summary: Mint a batch of unbound "blank card stock" credentials.
      tags: [credentials]
      x-sensitive: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IssueBlankBatchRequest"
      responses:
        "201":
          description: Created. Never cached.
          headers:
            Cache-Control:
              schema:
                type: string
                enum: [no-store]
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IssueBlankBatchResponse"
        default:
          $ref: "#/components/responses/ProblemResponse"
  /credentials/unbound-count:
    get:
      operationId: getUnboundCredentialCount
      summary: Count of unbound "blank card stock" credentials remaining.
      tags: [credentials]
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnboundCountResponse"
        default:
          $ref: "#/components/responses/ProblemResponse"
  /credentials/resolve:
    get:
      operationId: resolveCredential
      summary: Resolve a scanned token to its credential and subject, for binding a physical card during registration.
      tags: [credentials]
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResolvedCredential"
        default:
          $ref: "#/components/responses/ProblemResponse"
  /credentials/{id}/bind:
    post:
      operationId: bindCredential
      summary: Bind an unbound blank card to a subject.
      tags: [credentials]
      parameters:
        - $ref: "#/components/parameters/IDParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BindCredentialRequest"
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Credential"
        default:
          $ref: "#/components/responses/ProblemResponse"
  /credentials/{id}/reprint:
    post:
      operationId: reprintCredential
      summary: Recover and re-return the plaintext token of a device credential. Device tokens only.
      tags: [credentials]
      x-sensitive: true
      parameters:
        - $ref: "#/components/parameters/IDParam"
      responses:
        "200":
          description: OK. Never cached.
          headers:
            Cache-Control:
              schema:
                type: string
                enum: [no-store]
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IssuedCredential"
        default:
          $ref: "#/components/responses/ProblemResponse"
  /credentials/{id}/revoke:
    post:
      operationId: revokeCredential
      summary: Kill a token permanently, with no replacement minted.
      tags: [credentials]
      parameters:
        - $ref: "#/components/parameters/IDParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RevokeCredentialRequest"
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Credential"
        default:
          $ref: "#/components/responses/ProblemResponse"
  /credentials/{id}/reissue:
    post:
      operationId: reissueCredential
      summary: For a lost card — revoke the old one and mint a new one in the same transaction.
      tags: [credentials]
      x-sensitive: true
      parameters:
        - $ref: "#/components/parameters/IDParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReissueCredentialRequest"
      responses:
        "200":
          description: OK. Never cached.
          headers:
            Cache-Control:
              schema:
                type: string
                enum: [no-store]
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IssuedCredential"
        default:
          $ref: "#/components/responses/ProblemResponse"
  /credentials/{id}/history:
    get:
      operationId: getCredentialHistory
      summary: One credential's issuance history.
      tags: [credentials]
      parameters:
        - $ref: "#/components/parameters/IDParam"
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CredentialEventList"
        default:
          $ref: "#/components/responses/ProblemResponse"
components:
  schemas:
...
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        requestId:
          type: string
        extensions:
          type: object
          additionalProperties: true

    # --- Admin auth ---------------------------------------------------------
    AdminRole:
      type: string
      enum: [superadmin, admin, operator]
    Admin:
      type: object
      required: [id, email, fullName, role]
      properties:
        id: { type: string }
        email: { type: string }
        fullName: { type: string }
        role:
          $ref: "#/components/schemas/AdminRole"
...
    Category:
      type: object
      required: [id, name, requiresApproval, createdAt]
      properties:
        id: { type: string }
        name: { type: string }
        defaultLoanPeriodSeconds: { type: integer, format: int64 }
        requiresApproval: { type: boolean }
        createdAt: { type: string, format: date-time }
    CategoryList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Category"
    CreateCategoryRequest:
      type: object
      required: [name]
      properties:
        name: { type: string }
        defaultLoanPeriodSeconds: { type: integer, format: int64 }
        requiresApproval: { type: boolean }
    UpdateCategoryRequest:
      type: object
      required: [name]
      properties:
        name: { type: string }
        defaultLoanPeriodSeconds: { type: integer, format: int64 }
        requiresApproval: { type: boolean }

    # --- Identity ---------------------------------------------------------
...
    CredentialKind:
      type: string
      enum: [qr, code128, nfc, rfid, manual]
    SubjectType:
      type: string
      enum: [user, device]
    CredentialStatus:
      type: string
      enum: [active, revoked, lost]
    Credential:
      type: object
      required: [id, subjectType, kind, status, issueSeq, issuedAt, issuedBy, printedCount]
      properties:
        id: { type: string }
        subjectType:
          $ref: "#/components/schemas/SubjectType"
        subjectId:
          type: string
          description: Absent when the credential is unbound blank stock.
        kind:
          $ref: "#/components/schemas/CredentialKind"
        tokenPreview: { type: string }
        label: { type: string }
        status:
          $ref: "#/components/schemas/CredentialStatus"
        issueSeq: { type: integer }
        replacesId: { type: string }
        issuedAt: { type: string, format: date-time }
        issuedBy: { type: string }
        revokedAt: { type: string, format: date-time }
        revokedBy: { type: string }
        revokedReason: { type: string }
        printedCount: { type: integer }
        lastPrintedAt: { type: string, format: date-time }
    IssuedCredential:
      allOf:
        - $ref: "#/components/schemas/Credential"
        - type: object
          required: [token]
          properties:
            token:
              type: string
              description: Plaintext token. Returned exactly once, at the moment of issuance, reprint, or reissue — never retrievable again after this response.
    CredentialList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Credential"
    IssueCredentialRequest:
      type: object
      required: [subjectType, kind]
      properties:
        subjectType:
          $ref: "#/components/schemas/SubjectType"
        subjectId:
          type: string
          description: Omit with subjectType "user" to mint unbound blank card stock.
        kind:
          $ref: "#/components/schemas/CredentialKind"
        label: { type: string }
        manualToken:
          type: string
          description: Required, and used verbatim, only when kind is "manual".
    IssueBlankBatchRequest:
      type: object
      required: [count, kind]
      properties:
        count: { type: integer, minimum: 1 }
        kind:
          $ref: "#/components/schemas/CredentialKind"
    IssueBlankBatchResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/IssuedCredential"
    BindCredentialRequest:
      type: object
      required: [subjectId]
      properties:
        subjectId: { type: string }
    RevokeCredentialRequest:
      type: object
      required: [reason]
      properties:
        reason: { type: string }
    ReissueCredentialRequest:
      type: object
      required: [reason]
      properties:
        reason: { type: string }
        manualToken: { type: string }
    CredentialEvent:
      type: object
      required: [at, kind, actor]
      properties:
        at: { type: string, format: date-time }
        kind: { type: string }
        actor: { type: string }
        reason: { type: string }
    CredentialEventList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CredentialEvent"
    UnboundCountResponse:
      type: object
      required: [count]
      properties:
        count: { type: integer }
    ResolvedCredentialType:
      type: string
      description: >
        Resolve's subject discriminator. "unbound" is a successful
        resolution of a real, unbound blank card (INV-12) — not an error.
      enum: [user, device, unbound]
    ResolvedCredential:
      type: object
      required: [type, credentialId, credentialStatus, kind]
      properties:
        type:
          $ref: "#/components/schemas/ResolvedCredentialType"
        subjectId:
          type: string
          description: Absent when type is "unbound".
        credentialId: { type: string }
        credentialStatus:
          $ref: "#/components/schemas/CredentialStatus"
        kind:
          $ref: "#/components/schemas/CredentialKind"
  parameters:
    IDParam:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  responses:
    ProblemResponse:
      description: Error.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  securitySchemes:
    kioskToken:
      type: http
      scheme: bearer
      description: Long-lived opaque token issued when a kiosk is registered.
    adminSession:
      type: apiKey
      in: cookie
      name: hdms_session
    serviceToken:
      type: http
      scheme: bearer
      description: Used by imports, exports and scheduled jobs.